% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{golem_add_external_resources}
\alias{golem_add_external_resources}
\title{#' Add external Resources to the Application
#' 
#' This function is internally used to add external 
#' resources inside the Shiny application. 
#' 
#' @import shiny
#' @importFrom golem add_resource_path activate_js favicon bundle_resources
#' @noRd}
\usage{
golem_add_external_resources()
}
\description{
#' Add external Resources to the Application
#' 
#' This function is internally used to add external 
#' resources inside the Shiny application. 
#' 
#' @import shiny
#' @importFrom golem add_resource_path activate_js favicon bundle_resources
#' @noRd
}
