% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bib.R
\name{read_bib}
\alias{read_bib}
\title{read bibliographic data}
\usage{
read_bib(path, original = "Database")
}
\arguments{
\item{path}{file directory of your bibliographic data}

\item{original}{Source of the data."Database" means the data is from a database e.g. web of science. "Dois" means you collected dois by your self as a text file.}
}
\description{
read bibliographic data
}
