% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sci_down_orginal.R
\name{sci_down_orginal}
\alias{sci_down_orginal}
\title{Download sci_hub paper main function}
\usage{
sci_down_orginal(url, dois, dir = NULL)
}
\arguments{
\item{url}{sci url, if not offered by the user, sci-hub\.se will be used}

\item{dois}{dois of a list of papers you want to download}

\item{dir}{the file directory where you put the paper downloaded}
}
\description{
Download sci_hub paper main function
}
