% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sci_loc.R
\name{sci_loc}
\alias{sci_loc}
\title{Download sci_hub paper}
\usage{
sci_loc(path, input_url = NULL, original)
}
\arguments{
\item{path}{The file path of the input data}

\item{input_url}{sci url, if not offered by the user, sci-hub\.se will be used}

\item{original}{The source of your input data, from a database \(like web of science\) or from mendeley}
}
\description{
Download sci_hub paper
}
